pbr: Add openSUSE Leap support
Added openSUSE Leap support, which required overhauling a lot of code and configs.
Since openSUSE only has syslinux 4, I switched USB booting to grub2 as it was easier
to work with, and had to change up the iso layout and config to work properly for all
distros. Since openSUSE doesn't use authselect I had to figure out a way to generate
the proper pam configs with two different commands. I had to reconfigure how the
liveos was generated since openSUSE names their pkgs differently for some reason.
Changed up the grub.cfg file template also, to handle the different options properly.

+++ src/modules/backup.py
@@ -80,8 +80,11 @@ class Backup(object):
                 raise GeneralError()
         else:
             if (t == "nfs" or t == "cifs") and not rpmq(f"{t}-utils"):
-                self.log.error(f" Backup location type is set to {t}, but {t}-utils isn't installed, please install.")
-                raise ExistsError()
+                # On suse the pkg name is nfs-client.
+                if not rpmq(f"{t}-client"):
+                    self.log.error(f" Backup location type is set to {t}, but {t}-utils isn't installed, "
+                                   "please install.")
+                    raise ExistsError()
 
             # Mount bk_mount, which is where the backup archive
             # will be stored later on.
@@ -117,7 +120,7 @@ class Backup(object):
 
         # Copying the log file to the tmp_bk_dir here so that it gets written
         # potentially before the mount is un-mounted.
-        if self.tmp_bk_dir:
+        if self.tmp_bk_dir and exists("/var/log/pbr.log"):
             copyfile("/var/log/pbr.log", join(self.tmp_bk_dir, "pbr.log"))
 
         if self.mounted:
